+++ router/core/graph_server.go
@@ -133,6 +133,7 @@ func (b BuildGraphMuxOptions) IsBaseGraph() bool {
 type buildMultiGraphHandlerOptions struct {
 	baseMux               *chi.Mux
 	featureFlagConfigs    map[string]*nodev1.FeatureFlagRouterExecutionConfig
+	hashes                map[string]routerconfig.HashInfo
 	reloadPersistentState *ReloadPersistentState
 	currentGraphMuxes     map[string]*graphMux
 	changes               *routerconfig.Changes
@@ -347,7 +348,11 @@ func newGraphServer(routerCtx context.Context, r *Router, response *routerconfig
 
 	if needNewBaseGraphMux {
 		// build new base grap mux
-		s.logger.Debug("Will build a new base graph mux for new graph server")
+		s.logger.With(
+			zap.String("old_execution_config_hash", response.Hashes[""].OldHash), // empty on router boot
+			zap.String("new_execution_config_hash", response.Hashes[""].NewHash),
+		).Debug("Will build a new base graph mux for new graph server")
+
 		gm, err = s.buildGraphMux(BuildGraphMuxOptions{
 			RouterConfigVersion:   s.baseRouterConfigVersion,
 			EngineConfig:          response.Config.GetEngineConfig(),
@@ -361,7 +366,9 @@ func newGraphServer(routerCtx context.Context, r *Router, response *routerconfig
 			return nil, fmt.Errorf("failed to build base mux: %w", err)
 		}
 	} else {
-		s.logger.Debug("Will reuse old base graph mux for new graph server")
+		s.logger.With(zap.String("execution_config_hash", response.Hashes[""].NewHash)).
+			Debug("Will reuse old base graph mux for new graph server")
+
 		gm = mux
 		reusedMuxes = append(reusedMuxes, reusedGraphMux{key: "", mux: mux})
 	}
@@ -377,6 +384,7 @@ func newGraphServer(routerCtx context.Context, r *Router, response *routerconfig
 		reloadPersistentState: r.reloadPersistentState,
 		currentGraphMuxes:     currentMuxes,
 		changes:               response.Changes,
+		hashes:                response.Hashes,
 		defaultClientTLS:      defaultGRPCClientTLS,
 		perSubgraphTLS:        perSubgraphGRPCTLS,
 	})
@@ -581,8 +589,10 @@ func (s *graphServer) buildMultiGraphHandler(
 			if !hasChanged && !wasAdded {
 				oldGraphMux, exists := opts.currentGraphMuxes[featureFlagName]
 				if exists {
-					s.logger.Debug("will reuse feature flag mux for new graph server",
-						zap.String("flag", featureFlagName))
+					s.logger.Debug("Will reuse feature flag mux for new graph server",
+						zap.String("flag", featureFlagName),
+						zap.String("execution_config_hash", opts.hashes[featureFlagName].NewHash),
+					)
 					featureFlagToMux[featureFlagName] = oldGraphMux.mux
 					reused = append(reused, reusedGraphMux{key: featureFlagName, mux: oldGraphMux})
 					continue
@@ -590,8 +600,11 @@ func (s *graphServer) buildMultiGraphHandler(
 			}
 		}
 
-		s.logger.Debug("will create a new feature flag mux for new graph server",
-			zap.String("flag", featureFlagName))
+		s.logger.Debug("Will create a new feature flag mux for new graph server",
+			zap.String("flag", featureFlagName),
+			zap.String("old_execution_config_hash", opts.hashes[featureFlagName].OldHash), // empty on router boot
+			zap.String("new_execution_config_hash", opts.hashes[featureFlagName].NewHash),
+		)
 
 		gm, err := s.buildGraphMux(BuildGraphMuxOptions{
 			FeatureFlagName:       featureFlagName,